Separate into Exec and SSH from Analyze
